refactor(orders): split edit, add-detail, delivery, index and show
Extracts hooks (useEditOrder, useAddDetail, useDelivery) and presentational
components (accordion steps, variant field, orders table, order info card,
delivery lists) so every orders page and component stays under the size
limits. Refs #81.
